added objectallocation and processdata (single and parallel)

+++ GCB.Utility/Memory/Pressure.cs
@@ -2,7 +2,7 @@
 {
     public static class Pressure
     {
-        public static void PauseAfter(this int incrementalVariable, int nthOperation, int pauseTimeInMilliseconds)
+        public static void PauseAfter(this int incrementalVariable, int nthOperation, decimal pauseTimeInMilliseconds)
         {
             if(incrementalVariable % nthOperation == 0)
             {
